refactor: rename file to reverb_components.py
